Struct pte_flags::PteFlagsAarch64

pub struct PteFlagsAarch64(/* private fields */);

Available on AArch64 only.

Page table entry (PTE) flags on aarch64.

Note: items beginning with an underscore _ are not used in Theseus.

The designation of bits in each PageTableEntry is as such:

Bits [0:11] (inclusive) are reserved by hardware for access flags, cacheability flags, shareability flags, and TLB storage flags.

Bits [12:47] (inclusive) are reserved by hardware to hold the physical frame address.

Bits [48:49] (inclusive) are reserved as zero.

Bits [50:54] (inclusive) are reserved by hardware for more access flags.

Bits [55:58] (inclusive) are available for custom OS usage.

Bits [59:63] (inclusive) are reserved by hardware for extended access flags.

Assumed System Configuration

The system has been configured to use 48-bit physical addresses (aka “OA“s: Output Addresses).

The system has been configured to use only a single translation stage, Stage 1.

The MAIR index 0 has a Normal + Outer Shareable entry.

The MAIR index 1 has a “DEVICE nGnRE” entry.

Implementations

impl PteFlagsAarch64

pub const VALID: Self = _

If set, this page is currently “present” in memory.

If not set, this page is not in memory, which could mean one of several things:

The page is not mapped at all

The page has been temporarily paged/swapped to disk

The page is waiting to be mapped, i.e., for demand paging.

pub const PAGE_DESCRIPTOR: Self = _

If set, this represents a page descriptor.

If not set, this represents a block descriptor.

pub const _MAIR_INDEX_0: Self = _

Indicates the page’s cacheability is described by MAIR Index 0.

Theseus uses this index for “normal” memory.

pub const NORMAL_MEMORY: Self = _

This page maps “normal” memory, i.e., non-device memory.

Theseus uses MAIR_INDEX_0 for this type of memory.

pub const _MAIR_INDEX_1: Self = _

Indicates the page’s cacheability is described by MAIR Index 1.

Theseus uses this index for “device” memory.

pub const DEVICE_MEMORY: Self = _

This page maps device memory, i.e., memory-mapped I/O registers.

Theseus uses MAIR_INDEX_1 for this type of memory.

pub const _MAIR_INDEX_2: Self = _

Indicates the page’s cacheability is described by MAIR Index 2.

This is unused in Theseus.

pub const _MAIR_INDEX_3: Self = _

Indicates the page’s cacheability is described by MAIR Index 3.

This is unused in Theseus.

pub const _MAIR_INDEX_4: Self = _

Indicates the page’s cacheability is described by MAIR Index 4.

This is unused in Theseus.

pub const _MAIR_INDEX_5: Self = _

Indicates the page’s cacheability is described by MAIR Index 5.

This is unused in Theseus.

pub const _MAIR_INDEX_6: Self = _

Indicates the page’s cacheability is described by MAIR Index 6.

This is unused in Theseus.

pub const _MAIR_INDEX_7: Self = _

Indicates the page’s cacheability is described by MAIR Index 7.

This is unused in Theseus.

pub const _NON_SECURE_ACCESS: Self = _

If set, this page is accessible in both Secure and Non-Secure execution levels.

If not set, this page is accessible in only Secure execution levels.

This is unused in Theseus.

pub const _USER_ACCESSIBLE: Self = _

If set, userspace (unprivileged mode) can access this page.

If not set, only kernelspace (privileged mode) can access this page.

This is unused in Theseus because it is a single privilege level OS.

pub const READ_ONLY: Self = _

If set, this page is read-only.

If not set, this page is writable.

pub const _NON_SHAREABLE: Self = _

Indicates that only a single CPU core may access this page.

This is not used and not supported by Theseus; use Self::OUTER_SHAREABLE.

pub const OUTER_SHAREABLE: Self = _

Indicates that multiple CPUs from multiple clusters may access this page.

This is the default and the the only value used in Theseus (and most systems).

pub const _INNER_SHAREABLE: Self = _

Multiple cores from the same cluster can access this page. Indicates that multiple CPUs from only a single cluster may access this page.

This is not used and not supported by Theseus; use Self::OUTER_SHAREABLE.

pub const ACCESSED: Self = _

The hardware will set this bit when the page is accessed.

The OS can then clear this bit once it has acknowledged that the page was accessed, if it cares at all about this information.

On aarch64, an “Access Flag Fault” may be raised if this bit is not set when this page is first accessed and is trying to be cached in the TLB. This fault can only occur when the Access Flag bit is 0 and the flag is being managed by software.

Thus, Theseus currently always sets this bit by default.

pub const _NOT_GLOBAL: Self = _

If set, this page is mapped into only one or less than all address spaces, or is mapped differently across different address spaces, and thus be flushed out of the TLB when switching address spaces (page tables).

If not set, this page is mapped identically across all address spaces (all root page tables) and doesn’t need to be flushed out of the TLB when switching to another address space (page table).

Note: Theseus is a single address space system, so this flag makes no difference.

pub const _GUARDED_PAGE: Self = _

If set, this page is considered a “Guarded Page”, which can be used to protect against executing instructions that aren’t the intended target of a branch (e.g., with BTI instruction).

This is only available if FEAT_BTI is implemented; otherwise it is reserved as 0.

This is currently not used in Theseus.

pub const DIRTY: Self = _

The hardware will set this bit when the page has been written to.

The OS can then clear this bit once it has acknowledged that the page was written to, which is primarily useful for paging/swapping to disk.

pub const _CONTIGUOUS: Self = _

If set, this translation table entry is part of a set that is contiguous in memory with adjacent entries that also have this bit set.

If not set, this translation table entry is not contiguous in memory with entries that are adjancent to it.

This is useful for reducing TLB pressure because the TLB entries for multiple contiguous adjacent entries can be combined into one TLB entry.

This is currently not used in Theseus.

pub const _PRIV_EXEC_NEVER: Self = _

If set, this page is not executable by privileged levels (kernel).

If not set, this page is executable by privileged levels (kernel).

In Theseus, use Self::NOT_EXECUTABLE instead.

pub const _USER_EXEC_NEVER: Self = _

If set, this page is not executable by unprivileged levels (user).

If not set, this page is executable by unprivileged levels (user).

In Theseus, use Self::NOT_EXECUTABLE instead.

pub const NOT_EXECUTABLE: Self = _

If set, this page is not executable.

If not set, this page is executable.

pub const EXCLUSIVE: Self = _

See PteFlags::EXCLUSIVE. We use bit 55 because it is available for custom OS usage on both x86_64 and aarch64.

impl PteFlagsAarch64

pub const fn empty() -> Self

Get a flags value with all bits unset.

pub const fn all() -> Self

Get a flags value with all known bits set.

pub const fn bits(&self) -> u64

Get the underlying bits value.

The returned value is exactly the bits set in this flags value.

pub const fn from_bits(bits: u64) -> Option<Self>

Convert from a bits value.

This method will return None if any unknown bits are set.

pub const fn from_bits_truncate(bits: u64) -> Self

Convert from a bits value, unsetting any unknown bits.

pub const fn from_bits_retain(bits: u64) -> Self

Convert from a bits value exactly.

pub fn from_name(name: &str) -> Option<Self>

Get a flags value with the bits of a flag with the given name set.

This method will return None if name is empty or doesn’t correspond to any named flag.

pub const fn is_empty(&self) -> bool

Whether all bits in this flags value are unset.

pub const fn is_all(&self) -> bool

Whether all known bits in this flags value are set.

pub const fn intersects(&self, other: Self) -> bool

Whether any set bits in a source flags value are also set in a target flags value.

pub const fn contains(&self, other: Self) -> bool

Whether all set bits in a source flags value are also set in a target flags value.

pub fn insert(&mut self, other: Self)

The bitwise or (|) of the bits in two flags values.

pub fn remove(&mut self, other: Self)

The intersection of a source flags value with the complement of a target flags value (&!).

This method is not equivalent to self & !other when other has unknown bits set. remove won’t truncate other, but the ! operator will.

pub fn toggle(&mut self, other: Self)

The bitwise exclusive-or (^) of the bits in two flags values.

pub fn set(&mut self, other: Self, value: bool)

Call insert when value is true or remove when value is false.

pub const fn intersection(self, other: Self) -> Self

The bitwise and (&) of the bits in two flags values.

pub const fn union(self, other: Self) -> Self

The bitwise or (|) of the bits in two flags values.

pub const fn difference(self, other: Self) -> Self

The intersection of a source flags value with the complement of a target flags value (&!).

This method is not equivalent to self & !other when other has unknown bits set. difference won’t truncate other, but the ! operator will.

pub const fn symmetric_difference(self, other: Self) -> Self

The bitwise exclusive-or (^) of the bits in two flags values.

pub const fn complement(self) -> Self

The bitwise negation (!) of the bits in a flags value, truncating the result.

impl PteFlagsAarch64

pub const fn iter(&self) -> Iter<PteFlagsAarch64>

Yield a set of contained flags values.

Each yielded flags value will correspond to a defined named flag. Any unknown bits will be yielded together as a final flags value.

pub const fn iter_names(&self) -> IterNames<PteFlagsAarch64>

Yield a set of contained named flags values.

This method is like iter, except only yields bits in contained named flags. Any unknown bits, or bits not corresponding to a contained flag will not be yielded.

impl PteFlagsAarch64

Functions common to PTE flags on all architectures.

pub const fn new() -> Self

Returns a new PteFlagsAarch64 with the default value, in which:

NORMAL_MEMORY (not DEVICE_MEMORY) is set.

OUTER_SHAREABLE is set.

READ_ONLY is set.

PAGE_DESCRIPTOR is set.

ACCESSED is set.

NOT_GLOBAL is set.

the NOT_EXECUTABLE bits are set.

Note: the ACCESSED bit is set by default because Theseus currently doesn’t perform any paging/swapping of pages to disk, which is what this bit is typically used for. On aarch64, not setting this bit can cause an Access Flag Fault (which is useful only for software-managed LRU paging algorithms), so we just set that bit by default to avoid any faults that we don’t care about.

pub fn valid(self, enable: bool) -> Self

Returns a copy of this PteFlagsAarch64 with the VALID bit set or cleared.

If enable is true, this PTE will be considered “present” and “valid”, meaning that the mapping from this page to a physical frame is valid and that the translation of a virtual address in this page should succeed.

If enable is false, this PTE will be considered “invalid”, and any attempt to access it for translation purposes will cause a page fault.

pub fn writable(self, enable: bool) -> Self

Returns a copy of this PteFlagsAarch64 with the WRITABLE bit set or cleared.

If enable is true, this will be writable.

If enable is false, this will be read-only.

pub fn executable(self, enable: bool) -> Self

Returns a copy of this PteFlagsAarch64 with the NOT_EXECUTABLE bit cleared or set.

If enable is true, this page will be executable (NOT_EXECUTABLE will be cleared).

If enable is false, this page will be non-executable, which is the default (NOT_EXECUTABLE will be set).

pub fn device_memory(self, enable: bool) -> Self

Returns a copy of this PteFlagsAarch64 with the DEVICE_MEMORY bit set or cleared.

If enable is true, this will be non-cacheable device memory.

If enable is false, this will be “normal” memory, the default.

pub fn exclusive(self, enable: bool) -> Self

Returns a copy of this PteFlagsAarch64 with the EXCLUSIVE bit set or cleared.

If enable is true, this page will exclusively map its frame.

If enable is false, this page will NOT exclusively map its frame.

pub fn accessed(self, enable: bool) -> Self

Returns a copy of this PteFlagsAarch64 with the ACCESSED bit set or cleared.

Typically this is used to clear the ACCESSED bit, in order to indicate that the OS has “acknowledged” the fact that this page was accessed since the last time it checked.

If enable is true, this page will be marked as accessed.

If enable is false, this page will be marked as not accessed.

pub fn dirty(self, enable: bool) -> Self

Returns a copy of this PteFlagsAarch64 with the DIRTY bit set or cleared.

Typically this is used to clear the DIRTY bit, in order to indicate that the OS has “acknowledged” the fact that this page was written to since the last time it checked. This bit is typically set by the hardware.

If enable is true, this page will be marked as dirty.

If enable is false, this page will be marked as clean.

pub const fn is_valid(&self) -> bool

pub const fn is_writable(&self) -> bool

pub const fn is_executable(&self) -> bool

pub const fn is_device_memory(&self) -> bool

pub const fn is_dirty(&self) -> bool

pub const fn is_accessed(&self) -> bool

pub const fn is_exclusive(&self) -> bool

impl PteFlagsAarch64

Functions specific to aarch64 PTE flags only.

pub const MASKED_BITS_FOR_CONVERSION: PteFlagsAarch64 = _

The mask of bit ranges that cannot be handled by toggling, as they are not single bit values, but multi-bit selectors/indices.

Currently this includes:

The three bits [2:4] for MAIR index values.

The two bits [8:9] for shareability.

pub fn adjust_for_higher_level_pte(self) -> Self

Returns a copy of this PteFlagsAarch64 with its flags adjusted for use in a higher-level page table entry, e.g., P4, P3, P2.

Currently, on aarch64, this does the following:

Clears the NOT_EXECUTABLE bit, making it executable.

P4, P3, and P2 entries should never set NOT_EXECUTABLE, only the lowest-level P1 entry should.

Clears the EXCLUSIVE bit.

Currently, we do not use the EXCLUSIVE bit for P4, P3, or P2 entries, because another page table frame may re-use it (create another alias to it) without our page table implementation knowing about it.

Only P1-level PTEs can map a frame exclusively.

Sets the ACCESSED bit, since Theseus currently does not use it and aarch64 will throw an Access Flag Fault if it is not set.

Sets the PAGE_DESCRIPTOR bit, since Theseus currently does not use block descriptors on aarch64.

Sets the VALID bit, as every P4, P3, and P2 entry must be valid.

pub fn page_descriptor(self, enable: bool) -> Self

Returns a copy of this PteFlagsAarch64 with the PAGE_DESCRIPTOR bit set or cleared.

If enable is true, this will represent a page descriptor.

If enable is false, this will represent a block descriptor.

pub const fn is_page_descriptor(&self) -> bool